added Post request for likes and update the message
