+++ src/App.jsx
@@ -1,14 +1,12 @@
 import { Header } from './Components/Header'
-import { Form } from './Components/Form'
-import { List } from './Components/List'
+import { ThoughtsApp } from './Components/ThoughtsApp'
 import { Footer } from './Components/Footer'
 
 export const App = () => {
   return (
     <div>
       <Header />
-      <Form />
-      <List />
+      <ThoughtsApp />
       <Footer />
     </div>
   )
